+++ AppleALC/kern_resources.cpp
@@ -7195,18 +7195,12 @@ static const KextPatch patches17[] {
 	{ { &kextList[3], patchBuf128, patchBuf129, 12, 4 }, 13, KernelPatcher::KernelAny },
 	{ { &kextList[3], patchBuf130, patchBuf131, 12, 6 }, 13, KernelPatcher::KernelAny },
 };
-static const uint8_t patchBuf132[] { 0x0A, 0x00, 0xDE, 0x10, };
-static const uint8_t patchBuf133[] { 0x0D, 0x00, 0xDE, 0x10, };
-static const KextPatch patches18[] {
-	{ { &kextList[2], patchBuf132, patchBuf133, 4, 2 }, 10, KernelPatcher::KernelAny },
-};
 ControllerModInfo controllerMod[] {
 	{ "HD4600 controller", 0x8086, 0xC0C, nullptr, 0, ControllerModInfo::PlatformAny, patches13, 5 },
 	{ "X99 HDEF controller", 0x8086, 0x8D20, nullptr, 0, ControllerModInfo::PlatformAny, patches14, 1 },
 	{ "Intel HD4600 HDMI (ig-platform-id 0x0d220003)", 0x8086, 0x412, nullptr, 0, 220332035, patches15, 3 },
 	{ "Intel HD4000 HDMI (ig-platform-id 0x01660003)", 0x8086, 0x166, nullptr, 0, 23461891, patches16, 3 },
 	{ "Intel HD4000 HDMI (ig-platform-id 0x01660009)", 0x8086, 0x166, nullptr, 0, 23461897, patches17, 2 },
-	{ "Nvidia GT240 HDMI", 0x0, 0xBE4, nullptr, 0, ControllerModInfo::PlatformAny, patches18, 1 },
 };
 
-const size_t controllerModSize {6};
+const size_t controllerModSize {5};